Separated find method from tracker handler
