+++ src/handlers/tracker.ts
@@ -12,29 +12,21 @@ import {
   WidgetTracker
 } from '@jupyterlab/apputils';
 
-import {
-  CodeEditor,
-  CodeEditorWrapper,
-  IEditorServices
-} from '@jupyterlab/codeeditor';
-
-import { IConsoleTracker } from '@jupyterlab/console';
+import { CodeEditorWrapper, IEditorServices } from '@jupyterlab/codeeditor';
 
 import { PathExt } from '@jupyterlab/coreutils';
 
-import { IEditorTracker } from '@jupyterlab/fileeditor';
-
-import { INotebookTracker } from '@jupyterlab/notebook';
-
 import { textEditorIcon } from '@jupyterlab/ui-components';
 
-import { chain, each } from '@lumino/algorithm';
+import { each } from '@lumino/algorithm';
 
 import { Token } from '@lumino/coreutils';
 
 import { IDisposable } from '@lumino/disposable';
 
 import { Signal } from '@lumino/signaling';
+
+import { IDebuggerEditorFinder } from '../editor-finder';
 
 import { EditorHandler } from './editor';
 
@@ -59,10 +51,6 @@ export class TrackerHandler implements IDisposable {
   constructor(options: TrackerHandler.IOptions) {
     this._debuggerService = options.debuggerService;
     this._shell = options.shell;
-    this._notebookTracker = options.notebookTracker;
-    this._consoleTracker = options.consoleTracker;
-    this._editorTracker = options.editorTracker;
-
     this._readOnlyEditorFactory = new ReadOnlyEditorFactory({
       editorServices: options.editorServices
     });
@@ -73,6 +61,7 @@ export class TrackerHandler implements IDisposable {
       namespace: '@jupyterlab/debugger'
     });
 
+    this._editorFinder = options.editorFinder;
     this._onModelChanged();
     this._debuggerService.modelChanged.connect(this._onModelChanged, this);
   }
@@ -133,7 +122,7 @@ export class TrackerHandler implements IDisposable {
   ) {
     const debugSessionPath = this._debuggerService.session?.connection?.path;
     const source = frame?.source.path ?? null;
-    each(this._find(debugSessionPath, source), editor => {
+    each(this._editorFinder.find(debugSessionPath, source), editor => {
       requestAnimationFrame(() => {
         EditorHandler.showCurrentLine(editor, frame.line);
       });
@@ -151,7 +140,7 @@ export class TrackerHandler implements IDisposable {
     }
     const debugSessionPath = this._debuggerService.session.connection.path;
     const { content, mimeType, path } = source;
-    const results = this._find(debugSessionPath, path);
+    const results = this._editorFinder.find(debugSessionPath, path);
     if (results.next()) {
       return;
     }
@@ -183,149 +172,6 @@ export class TrackerHandler implements IDisposable {
       EditorHandler.showCurrentLine(editor, frame.line);
     }
   }
-
-  /**
-   * Find the editor for a source matching the current debug session
-   * by iterating through all the widgets in each of the notebook,
-   * console, file editor, and read-only file editor trackers.
-   * @param debugSessionPath The path for the current debug session.
-   * @param source The source to find.
-   */
-  private _find(debugSessionPath: string, source: string) {
-    return chain(
-      this._findInNotebooks(debugSessionPath, source),
-      this._findInConsoles(debugSessionPath, source),
-      this._findInEditors(debugSessionPath, source),
-      this._findInReadOnlyEditors(debugSessionPath, source)
-    );
-  }
-
-  /**
-   * Find the editor for a source matching the current debug session
-   * from the notebook tracker.
-   * @param debugSessionPath The path for the current debug session.
-   * @param source The source to find.
-   */
-  private _findInNotebooks(debugSessionPath: string, source: string) {
-    if (!this._notebookTracker) {
-      return [];
-    }
-    const editors: CodeEditor.IEditor[] = [];
-    this._notebookTracker.forEach(notebookPanel => {
-      const sessionContext = notebookPanel.sessionContext;
-
-      if (sessionContext.path !== debugSessionPath) {
-        return;
-      }
-
-      const notebook = notebookPanel.content;
-      notebook.mode = 'command';
-      const cells = notebookPanel.content.widgets;
-      cells.forEach((cell, i) => {
-        // check the event is for the correct cell
-        const code = cell.model.value.text;
-        const cellId = this._debuggerService.getCodeId(code);
-        if (source !== cellId) {
-          return;
-        }
-        notebook.activeCellIndex = i;
-        const rect = notebook.activeCell.inputArea.node.getBoundingClientRect();
-        notebook.scrollToPosition(rect.bottom, 45);
-        editors.push(cell.editor);
-        this._shell.activateById(notebookPanel.id);
-      });
-    });
-    return editors;
-  }
-
-  /**
-   * Find the editor for a source matching the current debug session
-   * from the console tracker.
-   * @param debugSessionPath The path for the current debug session.
-   * @param source The source to find.
-   */
-  private _findInConsoles(debugSessionPath: string, source: string) {
-    if (!this._consoleTracker) {
-      return [];
-    }
-    const editors: CodeEditor.IEditor[] = [];
-    this._consoleTracker.forEach(consoleWidget => {
-      const sessionContext = consoleWidget.sessionContext;
-
-      if (sessionContext.path !== debugSessionPath) {
-        return;
-      }
-
-      const cells = consoleWidget.console.cells;
-      each(cells, cell => {
-        const code = cell.model.value.text;
-        const codeId = this._debuggerService.getCodeId(code);
-        if (source !== codeId) {
-          return;
-        }
-        editors.push(cell.editor);
-        this._shell.activateById(consoleWidget.id);
-      });
-    });
-    return editors;
-  }
-
-  /**
-   * Find the editor for a source matching the current debug session
-   * from the editor tracker.
-   * @param debugSessionPath The path for the current debug session.
-   * @param source The source to find.
-   */
-  private _findInEditors(debugSessionPath: string, source: string) {
-    if (!this._editorTracker) {
-      return;
-    }
-    const editors: CodeEditor.IEditor[] = [];
-    this._editorTracker.forEach(doc => {
-      const fileEditor = doc.content;
-      if (debugSessionPath !== fileEditor.context.path) {
-        return;
-      }
-
-      const editor = fileEditor.editor;
-      if (!editor) {
-        return;
-      }
-
-      const code = editor.model.value.text;
-      const codeId = this._debuggerService.getCodeId(code);
-      if (source !== codeId) {
-        return;
-      }
-      editors.push(editor);
-      this._shell.activateById(doc.id);
-    });
-    return editors;
-  }
-
-  /**
-   * Find an editor for a source from the read-only editor tracker.
-   * @param source The source to find.
-   */
-  private _findInReadOnlyEditors(_: string, source: string) {
-    const editors: CodeEditor.IEditor[] = [];
-    this._readOnlyEditorTracker.forEach(widget => {
-      const editor = widget.content?.editor;
-      if (!editor) {
-        return;
-      }
-
-      const code = editor.model.value.text;
-      const codeId = this._debuggerService.getCodeId(code);
-      if (widget.title.caption !== source && source !== codeId) {
-        return;
-      }
-      editors.push(editor);
-      this._shell.activateById(widget.id);
-    });
-    return editors;
-  }
-
   private _debuggerService: IDebugger;
   private _debuggerModel: DebuggerModel;
   private _shell: JupyterFrontEnd.IShell;
@@ -333,9 +179,7 @@ export class TrackerHandler implements IDisposable {
   private _readOnlyEditorTracker: WidgetTracker<
     MainAreaWidget<CodeEditorWrapper>
   >;
-  private _notebookTracker: INotebookTracker | null;
-  private _consoleTracker: IConsoleTracker | null;
-  private _editorTracker: IEditorTracker | null;
+  private _editorFinder: IDebuggerEditorFinder | null;
 }
 
 /**
@@ -360,21 +204,10 @@ export namespace TrackerHandler {
      * The application shell.
      */
     shell: JupyterFrontEnd.IShell;
-
     /**
-     * An optional tracker for notebooks.
+     * An optional editor finder
      */
-    notebookTracker?: INotebookTracker;
-
-    /**
-     * An optional tracker for consoles.
-     */
-    consoleTracker?: IConsoleTracker;
-
-    /**
-     * An optional tracker for file editors.
-     */
-    editorTracker?: IEditorTracker;
+    editorFinder?: IDebuggerEditorFinder;
   }
 
   // TODO: move the interface and token below to token.ts?
